refactor(core): modularize codebase to match CLAUDE.md target architecture
This is a comprehensive refactoring that splits the monolithic main.rs (3474 lines)
into a proper modular structure following the architecture specified in CLAUDE.md:

**Module Structure:**
- src/main.rs (67 lines) - Minimal entry point with app initialization
- src/app.rs (2139 lines) - Heike struct, update loop, and core logic
- src/entry.rs (99 lines) - FileEntry struct and implementations
- src/style.rs (76 lines) - Theme enum and layout constants (replaces layout.rs)
- src/input.rs (3 lines) - Placeholder for future keyboard handling extraction

**New Directories:**
- src/state/ (4 files, 64 lines total)
  - mod.rs - Module re-exports
  - mode.rs - AppMode enum
  - clipboard.rs - ClipboardOp enum
  - search.rs - SearchResult and SearchOptions

- src/io/ (4 files, 507 lines total)
  - mod.rs - Module re-exports
  - directory.rs - Directory reading, fuzzy matching, binary detection
  - search.rs - Content search (PDF, ZIP, DOCX, XLSX, text files)
  - worker.rs - IoCommand, IoResult, async worker thread

- src/view/ (4 files, 808 lines total)
  - mod.rs - Module re-exports
  - preview.rs - All preview rendering (syntax, markdown, archives, etc.)
  - panels.rs - Placeholder for panel rendering extraction
  - modals.rs - Placeholder for modal rendering extraction

**Key Changes:**
- Removed old layout.rs (replaced by style.rs)
- Fixed deprecated egui API calls (screen_rect → viewport().inner_rect)
- Cleaned up unused imports
- All existing functionality preserved with zero behavior changes

**Benefits:**
- Clear separation of concerns (state, I/O, views, core logic)
- Much easier to locate and modify specific functionality
- Better testability with standalone modules
- Scalable architecture for future features
- Follows Rust best practices for module organization

Verified: Compiles successfully with only harmless warnings about unused constants.
