+++ src/app.rs
@@ -1,0 +1,2139 @@
+use crate::entry::FileEntry;
+use crate::io::{fuzzy_match, is_likely_binary, read_directory, spawn_worker, IoCommand, IoResult};
+use crate::state::{AppMode, ClipboardOp, SearchOptions};
+use crate::style::{self, Theme};
+use crate::view;
+
+use chrono::{DateTime, Local};
+use eframe::egui;
+use notify::{Event, RecursiveMode, Watcher};
+use std::collections::{HashMap, HashSet};
+use std::env;
+use std::fs;
+use std::path::PathBuf;
+use std::sync::mpsc::{channel, Receiver, Sender};
+use std::time::{Duration, Instant};
+use syntect::highlighting::ThemeSet;
+use syntect::parsing::SyntaxSet;
+
+pub struct Heike {
+    pub current_path: PathBuf,
+    pub history: Vec<PathBuf>,
+    pub history_index: usize,
+    pub all_entries: Vec<FileEntry>,
+    pub visible_entries: Vec<FileEntry>,
+    pub parent_entries: Vec<FileEntry>,
+    pub selected_index: Option<usize>,
+    pub multi_selection: HashSet<PathBuf>,
+    pub directory_selections: HashMap<PathBuf, usize>,
+    pub pending_selection_path: Option<PathBuf>,
+    pub mode: AppMode,
+    pub command_buffer: String,
+    pub focus_input: bool,
+    pub clipboard: HashSet<PathBuf>,
+    pub clipboard_op: Option<ClipboardOp>,
+    pub search_query: String,
+    pub search_options: SearchOptions,
+    pub search_in_progress: bool,
+    pub search_file_count: usize,
+    pub error_message: Option<(String, Instant)>,
+    pub info_message: Option<(String, Instant)>,
+    pub show_hidden: bool,
+    pub theme: Theme,
+    pub is_loading: bool,
+    pub last_g_press: Option<Instant>,
+    pub last_selection_change: Instant,
+    pub disable_autoscroll: bool,
+    pub panel_widths: [f32; 2],
+    pub dragging_divider: Option<usize>,
+    pub last_screen_size: egui::Vec2,
+    pub command_tx: Sender<IoCommand>,
+    pub result_rx: Receiver<IoResult>,
+    pub watcher: Option<Box<dyn Watcher>>,
+    pub watcher_rx: Receiver<Result<Event, notify::Error>>,
+    pub watched_path: Option<PathBuf>,
+    pub syntax_set: SyntaxSet,
+    pub theme_set: ThemeSet,
+}
+impl Heike {
+    pub fn new(ctx: egui::Context) -> Self {
+        let start_path = directories::UserDirs::new()
+            .map(|ud| ud.home_dir().to_path_buf())
+            .unwrap_or_else(|| env::current_dir().unwrap_or_default());
+
+        let (cmd_tx, res_rx) = spawn_worker(ctx.clone());
+        let (_watch_tx, watch_rx) = channel();
+
+        let mut app = Self {
+            current_path: start_path.clone(),
+            history: vec![start_path.clone()],
+            history_index: 0,
+            all_entries: Vec::new(),
+            visible_entries: Vec::new(),
+            parent_entries: Vec::new(),
+            selected_index: Some(0),
+            multi_selection: HashSet::new(),
+            directory_selections: HashMap::new(),
+            pending_selection_path: None,
+            mode: AppMode::Normal,
+            command_buffer: String::new(),
+            focus_input: false,
+            clipboard: HashSet::new(), // Init
+            clipboard_op: None,        // Init
+            search_query: String::new(),
+            search_options: SearchOptions::default(),
+            search_in_progress: false,
+            search_file_count: 0,
+            error_message: None,
+            info_message: None,
+            show_hidden: false,
+            theme: Theme::Dark,
+            is_loading: false,
+            last_g_press: None,
+            last_selection_change: Instant::now(),
+            disable_autoscroll: false,
+            panel_widths: [style::PARENT_DEFAULT, style::PREVIEW_DEFAULT],
+            dragging_divider: None,
+            last_screen_size: egui::Vec2::ZERO,
+            command_tx: cmd_tx,
+            result_rx: res_rx,
+            watcher: None,
+            watcher_rx: watch_rx,
+            watched_path: None,
+            syntax_set: SyntaxSet::load_defaults_newlines(),
+            theme_set: ThemeSet::load_defaults(),
+        };
+
+        app.request_refresh();
+        app
+    }
+
+    fn request_refresh(&mut self) {
+        self.is_loading = true;
+        self.error_message = None;
+        // Keep info message if it's fresh, or maybe clear it? Let's keep it for feedback.
+        let _ = self.command_tx.send(IoCommand::LoadDirectory(
+            self.current_path.clone(),
+            self.show_hidden,
+        ));
+        if let Some(parent) = self.current_path.parent() {
+            let _ = self.command_tx.send(IoCommand::LoadParent(
+                parent.to_path_buf(),
+                self.show_hidden,
+            ));
+        } else {
+            self.parent_entries.clear();
+        }
+    }
+
+    fn apply_filter(&mut self) {
+        // Save currently selected item path before filtering
+        let previously_selected = self
+            .selected_index
+            .and_then(|idx| self.visible_entries.get(idx))
+            .map(|e| e.path.clone());
+
+        if self.mode == AppMode::Filtering && !self.command_buffer.is_empty() {
+            let query = self.command_buffer.clone();
+            self.visible_entries = self
+                .all_entries
+                .iter()
+                .filter(|e| fuzzy_match(&e.name, &query))
+                .cloned()
+                .collect();
+        } else {
+            self.visible_entries = self.all_entries.clone();
+        }
+
+        // Restore selection to previously selected item if possible
+        if let Some(path) = previously_selected {
+            if let Some(idx) = self.visible_entries.iter().position(|e| e.path == path) {
+                self.selected_index = Some(idx);
+            } else if !self.visible_entries.is_empty() {
+                self.selected_index = Some(0);
+            } else {
+                self.selected_index = None;
+            }
+        } else if self.visible_entries.is_empty() {
+            self.selected_index = None;
+        } else if self.selected_index.is_none() {
+            self.selected_index = Some(0);
+        }
+        self.validate_selection();
+    }
+
+    fn setup_watcher(&mut self, ctx: &egui::Context) {
+        // Only setup if path changed
+        if self.watched_path.as_ref() == Some(&self.current_path) {
+            return;
+        }
+
+        // Get the channel sender for watcher events
+        let (tx, rx) = channel();
+        self.watcher_rx = rx;
+
+        // Create the watcher
+        let ctx_clone = ctx.clone();
+        match notify::recommended_watcher(move |res: Result<Event, notify::Error>| {
+            let _ = tx.send(res);
+            ctx_clone.request_repaint();
+        }) {
+            Ok(mut watcher) => {
+                // Watch the current directory
+                if let Err(e) = watcher.watch(&self.current_path, RecursiveMode::NonRecursive) {
+                    self.error_message =
+                        Some((format!("Failed to watch directory: {}", e), Instant::now()));
+                    self.watcher = None;
+                    self.watched_path = None;
+                } else {
+                    self.watcher = Some(Box::new(watcher));
+                    self.watched_path = Some(self.current_path.clone());
+                }
+            }
+            Err(e) => {
+                self.error_message =
+                    Some((format!("Failed to create watcher: {}", e), Instant::now()));
+                self.watcher = None;
+                self.watched_path = None;
+            }
+        }
+    }
+
+    fn process_watcher_events(&mut self) {
+        while let Ok(event_result) = self.watcher_rx.try_recv() {
+            match event_result {
+                Ok(_event) => {
+                    // File system changed, trigger refresh
+                    self.request_refresh();
+                }
+                Err(e) => {
+                    // Watcher error, but don't show it to avoid spam
+                    eprintln!("Watcher error: {}", e);
+                }
+            }
+        }
+    }
+
+    fn process_async_results(&mut self) {
+        while let Ok(result) = self.result_rx.try_recv() {
+            match result {
+                IoResult::DirectoryLoaded { path, entries } => {
+                    if path != self.current_path {
+                        continue;
+                    }
+
+                    self.all_entries = entries;
+                    self.is_loading = false;
+                    self.apply_filter();
+
+                    // If there's a pending selection path, find and select it
+                    if let Some(pending_path) = self.pending_selection_path.take() {
+                        if let Some(idx) = self
+                            .visible_entries
+                            .iter()
+                            .position(|e| e.path == pending_path)
+                        {
+                            self.selected_index = Some(idx);
+                        }
+                    }
+
+                    // Validate selection after loading
+                    if let Some(idx) = self.selected_index {
+                        if idx >= self.visible_entries.len() && !self.visible_entries.is_empty() {
+                            self.selected_index = Some(self.visible_entries.len() - 1);
+                        }
+                    }
+                }
+                IoResult::ParentLoaded(entries) => {
+                    self.parent_entries = entries;
+                }
+                IoResult::SearchCompleted(results) => {
+                    self.search_in_progress = false;
+                    let result_count = results.len();
+                    self.mode = AppMode::SearchResults {
+                        query: self.search_query.clone(),
+                        results,
+                        selected_index: 0,
+                    };
+                    self.info_message = Some((
+                        format!(
+                            "Found {} matches in {} files",
+                            result_count, self.search_file_count
+                        ),
+                        Instant::now(),
+                    ));
+                }
+                IoResult::SearchProgress(count) => {
+                    self.search_file_count = count;
+                }
+                IoResult::Error(msg) => {
+                    self.is_loading = false;
+                    self.search_in_progress = false;
+                    self.error_message = Some((msg, Instant::now()));
+                    self.all_entries.clear();
+                    self.visible_entries.clear();
+                }
+            }
+        }
+    }
+
+    // --- Navigation Logic ---
+
+    fn navigate_to(&mut self, path: PathBuf) {
+        if path.is_dir() {
+            // Save current selection before navigating away
+            if let Some(idx) = self.selected_index {
+                self.directory_selections
+                    .insert(self.current_path.clone(), idx);
+            }
+
+            self.current_path = path.clone();
+
+            if self.history_index < self.history.len() - 1 {
+                self.history.truncate(self.history_index + 1);
+            }
+            self.history.push(path);
+            self.history_index = self.history.len() - 1;
+
+            self.finish_navigation();
+        } else if let Err(e) = open::that(&path) {
+            self.error_message = Some((format!("Could not open file: {}", e), Instant::now()));
+        }
+    }
+
+    fn navigate_up(&mut self) {
+        if let Some(parent) = self.current_path.parent() {
+            // Save current selection before navigating up
+            if let Some(idx) = self.selected_index {
+                self.directory_selections
+                    .insert(self.current_path.clone(), idx);
+            }
+            self.navigate_to(parent.to_path_buf());
+        }
+    }
+
+    fn navigate_back(&mut self) {
+        if self.history_index == 0 {
+            return;
+        }
+
+        if let Some(idx) = self.selected_index {
+            self.directory_selections
+                .insert(self.current_path.clone(), idx);
+        }
+
+        let mut idx = self.history_index;
+        while idx > 0 {
+            idx -= 1;
+            let target = self.history[idx].clone();
+            if target.is_dir() {
+                self.history_index = idx;
+                self.current_path = target;
+                self.finish_navigation();
+                return;
+            } else {
+                self.history.remove(idx);
+                self.history_index -= 1;
+            }
+        }
+
+        self.error_message = Some(("Previous directory no longer exists".into(), Instant::now()));
+    }
+
+    fn navigate_forward(&mut self) {
+        if self.history_index >= self.history.len() - 1 {
+            return;
+        }
+
+        if let Some(idx) = self.selected_index {
+            self.directory_selections
+                .insert(self.current_path.clone(), idx);
+        }
+
+        let idx = self.history_index + 1;
+        loop {
+            if idx >= self.history.len() {
+                break;
+            }
+            let target = self.history[idx].clone();
+            if target.is_dir() {
+                self.history_index = idx;
+                self.current_path = target;
+                self.finish_navigation();
+                return;
+            }
+            self.history.remove(idx);
+        }
+
+        self.error_message = Some(("Next directory no longer exists".into(), Instant::now()));
+    }
+
+    fn finish_navigation(&mut self) {
+        self.command_buffer.clear();
+        self.mode = AppMode::Normal;
+        self.multi_selection.clear();
+        // Restore saved selection for this directory, or default to 0
+        self.selected_index = self
+            .directory_selections
+            .get(&self.current_path)
+            .copied()
+            .or(Some(0));
+        self.request_refresh();
+    }
+
+    // --- File Operations (Injected) ---
+
+    fn yank_selection(&mut self, op: ClipboardOp) {
+        self.clipboard.clear();
+        self.clipboard_op = Some(op);
+
+        if !self.multi_selection.is_empty() {
+            self.clipboard = self.multi_selection.clone();
+            self.mode = AppMode::Normal;
+            self.multi_selection.clear();
+        } else if let Some(idx) = self.selected_index {
+            if let Some(entry) = self.visible_entries.get(idx) {
+                self.clipboard.insert(entry.path.clone());
+            }
+        }
+
+        let op_text = if self.clipboard_op == Some(ClipboardOp::Copy) {
+            "Yanked"
+        } else {
+            "Cut"
+        };
+        self.info_message = Some((
+            format!("{} {} files", op_text, self.clipboard.len()),
+            Instant::now(),
+        ));
+    }
+
+    fn paste_clipboard(&mut self) {
+        if self.clipboard.is_empty() {
+            return;
+        }
+        let op = match self.clipboard_op {
+            Some(o) => o,
+            None => return,
+        };
+
+        let mut count = 0;
+        let mut errors = Vec::new();
+        let mut missing_paths = Vec::new();
+
+        for src in &self.clipboard {
+            if !src.exists() {
+                errors.push(format!("Source missing: {}", src.display()));
+                missing_paths.push(src.clone());
+                continue;
+            }
+
+            if let Some(name) = src.file_name() {
+                let dest = self.current_path.join(name);
+                if src.is_dir() {
+                    if op == ClipboardOp::Cut {
+                        if let Err(e) = fs::rename(src, &dest) {
+                            errors.push(format!("Move dir failed: {}", e));
+                        } else {
+                            count += 1;
+                        }
+                    } else {
+                        errors.push("Copying directories not supported in  Heike (lite)".into());
+                    }
+                } else if op == ClipboardOp::Copy {
+                    if let Err(e) = fs::copy(src, &dest) {
+                        errors.push(format!("Copy file failed: {}", e));
+                    } else {
+                        count += 1;
+                    }
+                } else if let Err(e) = fs::rename(src, &dest) {
+                    errors.push(format!("Move file failed: {}", e));
+                } else {
+                    count += 1;
+                }
+            }
+        }
+
+        for path in missing_paths {
+            self.clipboard.remove(&path);
+        }
+
+        if !errors.is_empty() {
+            self.error_message = Some((errors.join(" | "), Instant::now()));
+        } else {
+            self.info_message = Some((format!("Processed {} files", count), Instant::now()));
+        }
+
+        if op == ClipboardOp::Cut {
+            self.clipboard.clear();
+            self.clipboard_op = None;
+        }
+        self.request_refresh();
+    }
+
+    fn perform_delete(&mut self) {
+        let targets = if !self.multi_selection.is_empty() {
+            self.multi_selection.clone()
+        } else if let Some(idx) = self.selected_index {
+            if let Some(entry) = self.visible_entries.get(idx) {
+                HashSet::from([entry.path.clone()])
+            } else {
+                HashSet::new()
+            }
+        } else {
+            HashSet::new()
+        };
+
+        for path in targets {
+            if path.is_dir() {
+                let _ = fs::remove_dir_all(&path);
+            } else {
+                let _ = fs::remove_file(&path);
+            }
+        }
+
+        self.mode = AppMode::Normal;
+        self.multi_selection.clear();
+        self.request_refresh();
+        self.info_message = Some(("Items deleted".into(), Instant::now()));
+    }
+
+    fn perform_rename(&mut self) {
+        if let Some(idx) = self.selected_index {
+            if let Some(entry) = self.visible_entries.get(idx) {
+                let new_name = self.command_buffer.trim();
+                if !new_name.is_empty() {
+                    let new_path = entry.path.parent().unwrap().join(new_name);
+                    if let Err(e) = fs::rename(&entry.path, &new_path) {
+                        self.error_message =
+                            Some((format!("Rename failed: {}", e), Instant::now()));
+                    } else {
+                        self.info_message = Some(("Renamed successfully".into(), Instant::now()));
+                    }
+                }
+            }
+        }
+        self.mode = AppMode::Normal;
+        self.command_buffer.clear();
+        self.request_refresh();
+    }
+
+    // --- Selection Validation ---
+
+    fn validate_selection(&mut self) {
+        if let Some(idx) = self.selected_index {
+            if self.visible_entries.is_empty() {
+                self.selected_index = None;
+            } else if idx >= self.visible_entries.len() {
+                self.selected_index = Some(self.visible_entries.len() - 1);
+            }
+        }
+    }
+
+    // --- Drag and Drop Handling ---
+    // (Currently handled in the eframe::App update method)
+
+    fn render_divider(&mut self, ui: &mut egui::Ui, index: usize) {
+        let response = ui.allocate_response(ui.available_size(), egui::Sense::drag());
+
+        let color = if response.hovered() || response.dragged() {
+            ui.visuals().widgets.active.bg_fill
+        } else {
+            egui::Color32::from_gray(60)
+        };
+        ui.painter().rect_filled(response.rect, 0.0, color);
+
+        if response.hovered() || response.dragged() {
+            ui.ctx().set_cursor_icon(egui::CursorIcon::ResizeHorizontal);
+        }
+
+        if response.dragged() {
+            let delta = response.drag_delta().x;
+            match index {
+                0 => {
+                    self.panel_widths[0] =
+                        (self.panel_widths[0] + delta).clamp(style::PARENT_MIN, style::PARENT_MAX)
+                }
+                1 => {
+                    self.panel_widths[1] = (self.panel_widths[1] - delta)
+                        .clamp(style::PREVIEW_MIN, style::PREVIEW_MAX)
+                }
+                _ => {}
+            }
+        }
+    }
+
+    fn render_preview(
+        &self,
+        ui: &mut egui::Ui,
+        next_navigation: &std::cell::RefCell<Option<PathBuf>>,
+        pending_selection: &std::cell::RefCell<Option<PathBuf>>,
+    ) {
+        let idx = match self.selected_index {
+            Some(i) => i,
+            None => {
+                ui.centered_and_justified(|ui| {
+                    ui.label("No file selected");
+                });
+                return;
+            }
+        };
+        let entry = match self.visible_entries.get(idx) {
+            Some(e) => e,
+            None => return,
+        };
+
+        style::truncated_label(
+            ui,
+            egui::RichText::new(format!("{} {}", entry.get_icon(), entry.display_name())).heading(),
+        );
+        ui.add_space(5.0);
+        style::truncated_label(
+            ui,
+            format!("Size: {}", bytesize::ByteSize(entry.size)),
+        );
+        let datetime: DateTime<Local> = entry.modified.into();
+        ui.label(format!("Modified: {}", datetime.format("%Y-%m-%d %H:%M")));
+        ui.separator();
+
+        if entry.is_dir {
+            // Show directory contents in preview pane
+            if self.last_selection_change.elapsed() <= Duration::from_millis(200) {
+                ui.centered_and_justified(|ui| {
+                    ui.spinner();
+                });
+                return;
+            }
+
+            match read_directory(&entry.path, self.show_hidden) {
+                Ok(entries) => {
+                    let accent = egui::Color32::from_rgb(120, 180, 255);
+                    let highlighted_index = self.directory_selections.get(&entry.path).copied();
+
+                    egui::ScrollArea::vertical()
+                        .id_salt("preview_dir")
+                        .auto_shrink([false, false])
+                        .max_height(ui.available_height())
+                        .show(ui, |ui| {
+                            ui.set_max_width(ui.available_width());
+                            let default_color = ui.visuals().text_color();
+                            use egui_extras::{Column, TableBuilder};
+                            TableBuilder::new(ui)
+                                .striped(true)
+                                .resizable(false)
+                                .cell_layout(egui::Layout::left_to_right(egui::Align::Center))
+                                .column(Column::auto().at_least(30.0))
+                                .column(Column::remainder().clip(true))
+                                .body(|body| {
+                                    body.rows(24.0, entries.len(), |mut row| {
+                                        let row_index = row.index();
+                                        let preview_entry = &entries[row_index];
+                                        let is_highlighted = highlighted_index == Some(row_index);
+                                        let text_color = if is_highlighted || preview_entry.is_dir {
+                                            accent
+                                        } else {
+                                            default_color
+                                        };
+                                        row.col(|ui| {
+                                            ui.label(
+                                                egui::RichText::new(preview_entry.get_icon())
+                                                    .size(14.0)
+                                                    .color(text_color),
+                                            );
+                                        });
+                                        row.col(|ui| {
+                                            let response = style::truncated_label_with_sense(
+                                                ui,
+                                                egui::RichText::new(preview_entry.display_name())
+                                                    .color(text_color),
+                                                egui::Sense::click(),
+                                            );
+                                            if response.clicked() {
+                                                // Navigate to the directory being previewed (the currently selected item)
+                                                // and set the clicked item to be selected after navigation
+                                                *next_navigation.borrow_mut() =
+                                                    Some(entry.path.clone());
+                                                *pending_selection.borrow_mut() =
+                                                    Some(preview_entry.path.clone());
+                                            }
+                                        });
+                                    });
+                                });
+                        });
+                }
+                Err(e) => {
+                    ui.centered_and_justified(|ui| {
+                        ui.colored_label(
+                            egui::Color32::RED,
+                            format!("Cannot read directory: {}", e),
+                        );
+                    });
+                }
+            }
+            return;
+        }
+        if self.last_selection_change.elapsed() <= Duration::from_millis(200) {
+            ui.centered_and_justified(|ui| {
+                ui.spinner();
+            });
+            return;
+        }
+
+        // Image preview
+        if matches!(
+            entry.extension.as_str(),
+            "png" | "jpg" | "jpeg" | "gif" | "webp" | "bmp" | "svg" | "ico"
+        ) {
+            let uri = format!("file://{}", entry.path.display());
+            egui::ScrollArea::vertical()
+                .id_salt("preview_img")
+                .auto_shrink([false, false])
+                .max_height(ui.available_height())
+                .show(ui, |ui| {
+                    ui.set_max_width(ui.available_width());
+                    let available = ui.available_size();
+                    ui.add(
+                        egui::Image::new(uri)
+                            .max_width(available.x)
+                            .max_height(available.y - 100.0)
+                            .maintain_aspect_ratio(true)
+                            .shrink_to_fit(),
+                    );
+                });
+            return;
+        }
+
+        // Markdown preview
+        if matches!(entry.extension.as_str(), "md" | "markdown") {
+            view::preview::render_markdown_preview(ui, entry);
+            return;
+        }
+
+        // Archive preview
+        if matches!(
+            entry.extension.as_str(),
+            "zip" | "tar" | "gz" | "tgz" | "bz2" | "xz"
+        ) {
+            view::preview::render_archive_preview(ui, entry);
+            return;
+        }
+
+        // Audio metadata preview
+        if matches!(
+            entry.extension.as_str(),
+            "mp3" | "flac" | "ogg" | "m4a" | "wav"
+        ) {
+            view::preview::render_audio_metadata(ui, entry);
+            return;
+        }
+
+        // PDF preview
+        if matches!(entry.extension.as_str(), "pdf") {
+            view::preview::render_pdf_preview(ui, entry);
+            return;
+        }
+
+        // Word document preview
+        if matches!(entry.extension.as_str(), "docx" | "doc") {
+            view::preview::render_docx_preview(ui, entry);
+            return;
+        }
+
+        // Excel spreadsheet preview
+        if matches!(entry.extension.as_str(), "xlsx" | "xls") {
+            view::preview::render_xlsx_preview(ui, entry);
+            return;
+        }
+
+        // Code/text files with syntax highlighting
+        let text_extensions = [
+            "rs",
+            "py",
+            "js",
+            "ts",
+            "jsx",
+            "tsx",
+            "c",
+            "cpp",
+            "h",
+            "hpp",
+            "java",
+            "go",
+            "rb",
+            "php",
+            "swift",
+            "kt",
+            "scala",
+            "sh",
+            "bash",
+            "zsh",
+            "fish",
+            "ps1",
+            "bat",
+            "cmd",
+            "html",
+            "css",
+            "scss",
+            "sass",
+            "less",
+            "xml",
+            "yaml",
+            "yml",
+            "toml",
+            "json",
+            "ini",
+            "cfg",
+            "txt",
+            "log",
+            "conf",
+            "config",
+            "env",
+            "gitignore",
+            "dockerignore",
+            "editorconfig",
+            "sql",
+            "r",
+            "lua",
+            "vim",
+            "el",
+            "clj",
+            "ex",
+            "exs",
+            "erl",
+            "hrl",
+            "hs",
+            "ml",
+            "fs",
+            "cs",
+            "vb",
+            "pl",
+            "pm",
+            "t",
+            "asm",
+            "s",
+            "d",
+            "diff",
+            "patch",
+            "mak",
+            "makefile",
+            "cmake",
+            "gradle",
+            "properties",
+            "prefs",
+            "plist",
+            "nix",
+            "lisp",
+            "scm",
+            "rkt",
+            "proto",
+            "thrift",
+            "graphql",
+            "gql",
+            "vue",
+            "svelte",
+            "astro",
+            "dart",
+            "nim",
+            "zig",
+            "v",
+            "vala",
+            "cr",
+            "rst",
+            "adoc",
+            "tex",
+            "bib",
+            "lock",
+        ];
+
+        let check_as_text = text_extensions.contains(&entry.extension.as_str())
+            || entry.extension.is_empty()
+            || entry.name.starts_with('.'); // Hidden config files often have no extension
+
+        if check_as_text {
+            if entry.size > style::MAX_PREVIEW_SIZE {
+                view::preview::render_large_file_message(ui, entry);
+                return;
+            }
+
+            if !is_likely_binary(&entry.path) {
+                view::preview::render_syntax_highlighted(ui, entry, &self.syntax_set, &self.theme_set, self.theme);
+                return;
+            }
+        }
+
+        // Binary file - show info instead of auto-loading hex
+        ui.centered_and_justified(|ui| {
+            ui.vertical_centered(|ui| {
+                ui.add_space(20.0);
+                ui.label(egui::RichText::new("📦 Binary File").size(18.0));
+                ui.add_space(10.0);
+                ui.label("Preview not available for this file type");
+                ui.add_space(5.0);
+                ui.label(format!("Extension: .{}", entry.extension));
+            });
+        });
+    }
+
+    // --- Drag and Drop Handling ---
+
+    fn handle_dropped_files(&mut self, dropped_files: &[egui::DroppedFile]) {
+        let mut count = 0;
+        let mut errors = Vec::new();
+
+        for file in dropped_files {
+            if let Some(path) = &file.path {
+                let dest = self.current_path.join(path.file_name().unwrap_or_default());
+
+                // Copy the dropped file to current directory
+                if path.is_dir() {
+                    errors.push("Copying directories not supported".into());
+                } else {
+                    match fs::copy(path, &dest) {
+                        Ok(_) => count += 1,
+                        Err(e) => errors.push(format!("Copy failed: {}", e)),
+                    }
+                }
+            }
+        }
+
+        if !errors.is_empty() {
+            self.error_message = Some((errors.join(" | "), Instant::now()));
+        } else if count > 0 {
+            self.info_message = Some((format!("Copied {} file(s)", count), Instant::now()));
+        }
+
+        if count > 0 {
+            self.request_refresh();
+        }
+    }
+
+    // --- Input Handling ---
+
+    fn handle_input(&mut self, ctx: &egui::Context) {
+        // 1. Modal Inputs (Command, Filter, Rename, SearchInput)
+        if matches!(
+            self.mode,
+            AppMode::Command | AppMode::Filtering | AppMode::Rename | AppMode::SearchInput
+        ) {
+            if ctx.input(|i| i.key_pressed(egui::Key::Enter)) {
+                match self.mode {
+                    AppMode::Rename => self.perform_rename(),
+                    AppMode::Command => self.execute_command(ctx),
+                    AppMode::Filtering => {
+                        // Finalize search and allow navigation in filtered results
+                        self.mode = AppMode::Normal;
+                        // Keep the filtered results
+                    }
+                    AppMode::SearchInput => {
+                        // Start search
+                        if !self.search_query.is_empty() {
+                            self.search_in_progress = true;
+                            self.search_file_count = 0;
+                            let _ = self.command_tx.send(IoCommand::SearchContent {
+                                query: self.search_query.clone(),
+                                root_path: self.current_path.clone(),
+                                options: self.search_options.clone(),
+                            });
+                        }
+                        self.mode = AppMode::Normal;
+                    }
+                    _ => {}
+                }
+            }
+            if self.mode == AppMode::Filtering && !ctx.input(|i| i.pointer.any_pressed()) {
+                // Implicitly handled
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
+                self.mode = AppMode::Normal;
+                self.command_buffer.clear();
+                self.apply_filter();
+            }
+            return;
+        }
+
+        // 2. Confirmation Modals
+        if self.mode == AppMode::DeleteConfirm {
+            if ctx.input(|i| i.key_pressed(egui::Key::Y) || i.key_pressed(egui::Key::Enter)) {
+                self.perform_delete();
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::N) || i.key_pressed(egui::Key::Escape)) {
+                self.mode = AppMode::Normal;
+            }
+            return;
+        }
+
+        if self.mode == AppMode::Help {
+            if ctx.input(|i| {
+                i.key_pressed(egui::Key::Escape)
+                    || i.key_pressed(egui::Key::Q)
+                    || i.key_pressed(egui::Key::Questionmark)
+            }) {
+                self.mode = AppMode::Normal;
+            }
+            return;
+        }
+
+        // Handle SearchResults mode navigation
+        if let AppMode::SearchResults {
+            query: ref current_query,
+            ref results,
+            ref mut selected_index,
+        } = self.mode
+        {
+            if ctx.input(|i| i.key_pressed(egui::Key::S) && i.modifiers.shift) {
+                self.search_query = current_query.clone();
+                self.search_in_progress = false;
+                self.search_file_count = 0;
+                self.mode = AppMode::SearchInput;
+                self.focus_input = true;
+                return;
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
+                self.mode = AppMode::Normal;
+                return;
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::N) && !i.modifiers.shift) {
+                if !results.is_empty() {
+                    *selected_index = (*selected_index + 1) % results.len();
+                }
+                return;
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::N) && i.modifiers.shift) {
+                if !results.is_empty() {
+                    *selected_index = if *selected_index == 0 {
+                        results.len() - 1
+                    } else {
+                        *selected_index - 1
+                    };
+                }
+                return;
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::Enter)) {
+                // Open the file at the match location
+                if let Some(result) = results.get(*selected_index) {
+                    if result.file_path.is_file() {
+                        let _ = open::that(&result.file_path);
+                    }
+                }
+                return;
+            }
+            // Allow other navigation within search results
+            if ctx.input(|i| i.key_pressed(egui::Key::ArrowDown) || i.key_pressed(egui::Key::J)) {
+                if !results.is_empty() {
+                    *selected_index = (*selected_index + 1) % results.len();
+                }
+                return;
+            }
+            if ctx.input(|i| i.key_pressed(egui::Key::ArrowUp) || i.key_pressed(egui::Key::K)) {
+                if !results.is_empty() {
+                    *selected_index = if *selected_index == 0 {
+                        results.len() - 1
+                    } else {
+                        *selected_index - 1
+                    };
+                }
+                return;
+            }
+            return; // Don't process other keys in search results mode
+        }
+
+        if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
+            self.mode = AppMode::Normal;
+            self.command_buffer.clear();
+            self.multi_selection.clear();
+            self.apply_filter();
+            return;
+        }
+
+        // 3. Global History keys
+        if ctx.input(|i| i.modifiers.alt && i.key_pressed(egui::Key::ArrowLeft)) {
+            self.navigate_back();
+            return;
+        }
+        if ctx.input(|i| i.modifiers.alt && i.key_pressed(egui::Key::ArrowRight)) {
+            self.navigate_forward();
+            return;
+        }
+
+        // 4. Normal Mode Triggers
+        if ctx.input(|i| i.key_pressed(egui::Key::Colon)) {
+            self.mode = AppMode::Command;
+            self.focus_input = true;
+            self.command_buffer.clear();
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::Slash)) {
+            self.mode = AppMode::Filtering;
+            self.focus_input = true;
+            self.command_buffer.clear();
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::Period)) {
+            self.show_hidden = !self.show_hidden;
+            self.request_refresh();
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::Questionmark)) {
+            self.mode = AppMode::Help;
+            return;
+        }
+        if self.mode == AppMode::Normal
+            && ctx.input(|i| i.key_pressed(egui::Key::V) && !i.modifiers.shift)
+        {
+            self.mode = AppMode::Visual;
+            if let Some(idx) = self.selected_index {
+                if let Some(entry) = self.visible_entries.get(idx) {
+                    self.multi_selection.insert(entry.path.clone());
+                }
+            }
+            return;
+        }
+        if self.mode == AppMode::Normal
+            && ctx.input(|i| i.key_pressed(egui::Key::V) && i.modifiers.shift)
+        {
+            // Shift+V: Enter visual mode and select all
+            self.mode = AppMode::Visual;
+            self.multi_selection.clear();
+            for entry in &self.visible_entries {
+                self.multi_selection.insert(entry.path.clone());
+            }
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::A) && i.modifiers.ctrl) {
+            // Ctrl+A: Select all
+            if self.mode != AppMode::Visual {
+                self.mode = AppMode::Visual;
+            }
+            self.multi_selection.clear();
+            for entry in &self.visible_entries {
+                self.multi_selection.insert(entry.path.clone());
+            }
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::Space)) {
+            // Space: Toggle selection of current item
+            if let Some(idx) = self.selected_index {
+                if let Some(entry) = self.visible_entries.get(idx) {
+                    if self.multi_selection.contains(&entry.path) {
+                        self.multi_selection.remove(&entry.path);
+                    } else {
+                        if self.mode != AppMode::Visual {
+                            self.mode = AppMode::Visual;
+                        }
+                        self.multi_selection.insert(entry.path.clone());
+                    }
+                }
+            }
+            return;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::S) && i.modifiers.shift) {
+            self.search_in_progress = false;
+            self.search_file_count = 0;
+            self.mode = AppMode::SearchInput;
+            self.focus_input = true;
+            return;
+        }
+
+        // 5. File Operation Triggers (Phase 6)
+        if ctx.input(|i| i.key_pressed(egui::Key::Y)) {
+            self.yank_selection(ClipboardOp::Copy);
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::X)) {
+            self.yank_selection(ClipboardOp::Cut);
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::P)) {
+            self.paste_clipboard();
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::D)) {
+            self.mode = AppMode::DeleteConfirm;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::R)) {
+            if let Some(idx) = self.selected_index {
+                if let Some(entry) = self.visible_entries.get(idx) {
+                    self.command_buffer = entry.name.clone();
+                    self.mode = AppMode::Rename;
+                    self.focus_input = true;
+                }
+            }
+        }
+
+        // 6. Navigation (j/k/arrows)
+        if self.visible_entries.is_empty() {
+            if ctx.input(|i| {
+                i.key_pressed(egui::Key::Backspace)
+                    || i.key_pressed(egui::Key::H)
+                    || i.key_pressed(egui::Key::ArrowLeft)
+            }) {
+                self.navigate_up();
+            }
+            return;
+        }
+
+        let mut changed = false;
+        let max_idx = self.visible_entries.len() - 1;
+        let current = self.selected_index.unwrap_or(0);
+        let mut new_index = current;
+
+        if ctx.input(|i| i.key_pressed(egui::Key::ArrowDown) || i.key_pressed(egui::Key::J)) {
+            new_index = if current >= max_idx { 0 } else { current + 1 };
+            changed = true;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::ArrowUp) || i.key_pressed(egui::Key::K)) {
+            new_index = if current == 0 { max_idx } else { current - 1 };
+            changed = true;
+        }
+        if ctx.input(|i| {
+            i.key_pressed(egui::Key::Backspace)
+                || i.key_pressed(egui::Key::H)
+                || i.key_pressed(egui::Key::ArrowLeft)
+        }) {
+            self.navigate_up();
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::Enter)) {
+            if let Some(idx) = self.selected_index {
+                if let Some(entry) = self.visible_entries.get(idx) {
+                    let path = entry.path.clone();
+                    self.navigate_to(path);
+                }
+            }
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::L) || i.key_pressed(egui::Key::ArrowRight)) {
+            if let Some(idx) = self.selected_index {
+                if let Some(entry) = self.visible_entries.get(idx) {
+                    if entry.is_dir {
+                        let path = entry.path.clone();
+                        self.navigate_to(path);
+                    }
+                }
+            }
+        }
+
+        if ctx.input(|i| i.key_pressed(egui::Key::G) && i.modifiers.shift) {
+            new_index = max_idx;
+            changed = true;
+        }
+        if ctx.input(|i| i.key_pressed(egui::Key::G) && !i.modifiers.shift) {
+            let now = Instant::now();
+            if let Some(last) = self.last_g_press {
+                if now.duration_since(last) < Duration::from_millis(500) {
+                    new_index = 0;
+                    self.last_g_press = None;
+                    changed = true;
+                } else {
+                    self.last_g_press = Some(now);
+                }
+            } else {
+                self.last_g_press = Some(now);
+            }
+        }
+        if let Some(last) = self.last_g_press {
+            if Instant::now().duration_since(last) > Duration::from_millis(500) {
+                self.last_g_press = None;
+            }
+        }
+
+        if changed {
+            self.selected_index = Some(new_index);
+            self.last_selection_change = Instant::now();
+            self.disable_autoscroll = false; // Re-enable autoscroll on keyboard navigation
+            if self.mode == AppMode::Visual {
+                if let Some(entry) = self.visible_entries.get(new_index) {
+                    self.multi_selection.insert(entry.path.clone());
+                }
+            }
+        }
+    }
+
+    // --- Rendering Methods ---
+
+    fn render_parent_pane(
+        &self,
+        ui: &mut egui::Ui,
+        next_navigation: &std::cell::RefCell<Option<PathBuf>>,
+    ) {
+        ui.add_space(4.0);
+        ui.vertical_centered(|ui| {
+            ui.heading("Parent");
+        });
+        ui.separator();
+        let accent = egui::Color32::from_rgb(120, 180, 255);
+        let default_color = ui.visuals().text_color();
+
+        egui::ScrollArea::vertical()
+            .id_salt("parent_scroll")
+            .auto_shrink([false, false])
+            .max_height(ui.available_height())
+            .show(ui, |ui| {
+                ui.set_max_width(ui.available_width());
+                use egui_extras::{Column, TableBuilder};
+                TableBuilder::new(ui)
+                    .striped(true)
+                    .resizable(false)
+                    .cell_layout(egui::Layout::left_to_right(egui::Align::Center))
+                    .column(Column::auto().at_least(30.0))
+                    .column(Column::remainder().clip(true))
+                    .body(|body| {
+                        body.rows(24.0, self.parent_entries.len(), |mut row| {
+                            let entry = &self.parent_entries[row.index()];
+                            let is_active = entry.path == self.current_path;
+
+                            let icon_color = if is_active { accent } else { default_color };
+
+                            row.col(|ui| {
+                                ui.label(
+                                    egui::RichText::new(entry.get_icon())
+                                        .size(14.0)
+                                        .color(icon_color),
+                                );
+                            });
+                            row.col(|ui| {
+                                let text_color = if is_active { accent } else { default_color };
+                                let response = style::truncated_label_with_sense(
+                                    ui,
+                                    egui::RichText::new(entry.display_name()).color(text_color),
+                                    egui::Sense::click(),
+                                );
+                                if response.clicked() {
+                                    // Navigate to the clicked directory in the parent pane
+                                    *next_navigation.borrow_mut() = Some(entry.path.clone());
+                                }
+                            });
+                        });
+                    });
+            });
+    }
+
+    fn render_current_pane(
+        &mut self,
+        ui: &mut egui::Ui,
+        next_navigation: &std::cell::RefCell<Option<PathBuf>>,
+        next_selection: &std::cell::RefCell<Option<usize>>,
+        context_action: &std::cell::RefCell<Option<Box<dyn FnOnce(&mut Self)>>>,
+        ctx: &egui::Context,
+    ) {
+        // Detect manual scrolling in the central panel
+        if ui.ui_contains_pointer()
+            && ctx.input(|i| {
+                i.smooth_scroll_delta != egui::Vec2::ZERO || i.raw_scroll_delta != egui::Vec2::ZERO
+            })
+        {
+            self.disable_autoscroll = true;
+        }
+
+        egui::ScrollArea::vertical()
+            .id_salt("current_scroll")
+            .auto_shrink([false, false])
+            .show(ui, |ui| {
+                use egui_extras::{Column, TableBuilder};
+                let mut table = TableBuilder::new(ui)
+                    .striped(true)
+                    .resizable(false)
+                    .cell_layout(egui::Layout::left_to_right(egui::Align::Center))
+                    .column(Column::initial(30.0))
+                    .column(Column::remainder().clip(true));
+
+                // Only scroll to selected row if autoscroll is not disabled
+                if !self.disable_autoscroll {
+                    if let Some(idx) = self.selected_index {
+                        table = table.scroll_to_row(idx, None);
+                    }
+                }
+
+                table
+                    .header(20.0, |mut header| {
+                        header.col(|ui| {
+                            ui.label("");
+                        });
+                        header.col(|ui| {
+                            ui.label("Name");
+                        });
+                    })
+                    .body(|body| {
+                        body.rows(24.0, self.visible_entries.len(), |mut row| {
+                            let row_index = row.index();
+                            let entry = &self.visible_entries[row_index];
+                            let is_focused = self.selected_index == Some(row_index);
+                            let is_multi_selected = self.multi_selection.contains(&entry.path);
+                            let is_cut = self.clipboard_op == Some(ClipboardOp::Cut)
+                                && self.clipboard.contains(&entry.path);
+
+                            if is_multi_selected || is_focused {
+                                row.set_selected(true);
+                            }
+
+                            // Icon column
+                            row.col(|ui| {
+                                ui.label(egui::RichText::new(entry.get_icon()).size(14.0));
+                            });
+
+                            // Name column with context menu
+                            row.col(|ui| {
+                                let mut text = egui::RichText::new(entry.display_name());
+                                if is_multi_selected {
+                                    text = text.color(egui::Color32::LIGHT_BLUE);
+                                } else if is_cut {
+                                    text = text.color(egui::Color32::from_white_alpha(100));
+                                // Dimmed
+                                } else if entry.is_dir {
+                                    text = text.color(egui::Color32::from_rgb(120, 180, 255));
+                                // Subtle blue for directories
+                                } else {
+                                    // Keep default text color for files
+                                }
+
+                                let response = style::truncated_label_with_sense(
+                                    ui,
+                                    text,
+                                    egui::Sense::click(),
+                                );
+
+                                // Single click for selection only
+                                if response.clicked() {
+                                    *next_selection.borrow_mut() = Some(row_index);
+                                }
+
+                                // Double click to open/navigate
+                                if response.double_clicked() {
+                                    if let Some(entry) = self.visible_entries.get(row_index) {
+                                        *next_navigation.borrow_mut() = Some(entry.path.clone());
+                                    }
+                                }
+
+                                // Context menu on right-click
+                                let entry_clone = entry.clone();
+                                response.context_menu(|ui| {
+                                    if ui.button("📂 Open").clicked() {
+                                        if entry_clone.is_dir {
+                                            *next_navigation.borrow_mut() =
+                                                Some(entry_clone.path.clone());
+                                        } else {
+                                            let _ = open::that(&entry_clone.path);
+                                        }
+                                        ui.close();
+                                    }
+
+                                    ui.separator();
+
+                                    if ui.button("📋 Copy (y)").clicked() {
+                                        let path = entry_clone.path.clone();
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(move |app: &mut Self| {
+                                                app.clipboard.clear();
+                                                app.clipboard.insert(path);
+                                                app.clipboard_op = Some(ClipboardOp::Copy);
+                                                app.info_message =
+                                                    Some(("Copied 1 file".into(), Instant::now()));
+                                            }));
+                                        ui.close();
+                                    }
+
+                                    if ui.button("✂️ Cut (x)").clicked() {
+                                        let path = entry_clone.path.clone();
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(move |app: &mut Self| {
+                                                app.clipboard.clear();
+                                                app.clipboard.insert(path);
+                                                app.clipboard_op = Some(ClipboardOp::Cut);
+                                                app.info_message =
+                                                    Some(("Cut 1 file".into(), Instant::now()));
+                                            }));
+                                        ui.close();
+                                    }
+
+                                    if ui.button("📥 Paste (p)").clicked() {
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(|app: &mut Self| {
+                                                app.paste_clipboard();
+                                            }));
+                                        ui.close();
+                                    }
+
+                                    ui.separator();
+
+                                    if ui.button("✏️ Rename (r)").clicked() {
+                                        *next_selection.borrow_mut() = Some(row_index);
+                                        let name = entry_clone.name.clone();
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(move |app: &mut Self| {
+                                                app.command_buffer = name;
+                                                app.mode = AppMode::Rename;
+                                                app.focus_input = true;
+                                            }));
+                                        ui.close();
+                                    }
+
+                                    if ui.button("🗑️ Delete (d)").clicked() {
+                                        *next_selection.borrow_mut() = Some(row_index);
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(|app: &mut Self| {
+                                                app.mode = AppMode::DeleteConfirm;
+                                            }));
+                                        ui.close();
+                                    }
+
+                                    ui.separator();
+
+                                    if ui.button("ℹ️ Properties").clicked() {
+                                        let size = entry_clone.size;
+                                        let modified = entry_clone.modified;
+                                        let is_dir = entry_clone.is_dir;
+                                        *context_action.borrow_mut() =
+                                            Some(Box::new(move |app: &mut Self| {
+                                                app.info_message =
+                                                    Some((
+                                                        format!(
+                                            "{} | {} | Modified: {}",
+                                            if is_dir { "Directory" } else { "File" },
+                                            bytesize::ByteSize(size),
+                                            chrono::DateTime::<chrono::Local>::from(modified)
+                                                .format("%Y-%m-%d %H:%M")
+                                        ),
+                                                        Instant::now(),
+                                                    ));
+                                            }));
+                                        ui.close();
+                                    }
+                                });
+                            });
+                        });
+                    });
+            });
+    }
+
+    // Note: execute_command method is missing but needed for handle_input to compile
+    fn execute_command(&mut self, _ctx: &egui::Context) {
+        let parts: Vec<&str> = self.command_buffer.trim().split_whitespace().collect();
+        if parts.is_empty() {
+            self.mode = AppMode::Normal;
+            self.command_buffer.clear();
+            return;
+        }
+
+        match parts[0] {
+            "q" | "quit" => {
+                std::process::exit(0);
+            }
+            "mkdir" => {
+                if parts.len() < 2 {
+                    self.error_message = Some(("Usage: mkdir <name>".into(), Instant::now()));
+                } else {
+                    let dir_name = parts[1..].join(" ");
+                    let new_dir = self.current_path.join(&dir_name);
+                    match fs::create_dir(&new_dir) {
+                        Ok(_) => {
+                            self.info_message =
+                                Some((format!("Created directory: {}", dir_name), Instant::now()));
+                            self.request_refresh();
+                        }
+                        Err(e) => {
+                            self.error_message =
+                                Some((format!("Failed to create directory: {}", e), Instant::now()));
+                        }
+                    }
+                }
+            }
+            "touch" => {
+                if parts.len() < 2 {
+                    self.error_message = Some(("Usage: touch <filename>".into(), Instant::now()));
+                } else {
+                    let file_name = parts[1..].join(" ");
+                    let new_file = self.current_path.join(&file_name);
+                    match fs::File::create(&new_file) {
+                        Ok(_) => {
+                            self.info_message =
+                                Some((format!("Created file: {}", file_name), Instant::now()));
+                            self.request_refresh();
+                        }
+                        Err(e) => {
+                            self.error_message =
+                                Some((format!("Failed to create file: {}", e), Instant::now()));
+                        }
+                    }
+                }
+            }
+            _ => {
+                self.error_message =
+                    Some((format!("Unknown command: {}", parts[0]), Instant::now()));
+            }
+        }
+
+        self.mode = AppMode::Normal;
+        self.command_buffer.clear();
+    }
+}
+
+impl eframe::App for Heike {
+    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
+        // Apply theme
+        match self.theme {
+            Theme::Light => ctx.set_visuals(egui::Visuals::light()),
+            Theme::Dark => ctx.set_visuals(egui::Visuals::dark()),
+        }
+
+        // Auto-dismiss old messages
+        if let Some((_, time)) = &self.error_message {
+            if time.elapsed() > Duration::from_secs(style::MESSAGE_TIMEOUT_SECS) {
+                self.error_message = None;
+            }
+        }
+        if let Some((_, time)) = &self.info_message {
+            if time.elapsed() > Duration::from_secs(style::MESSAGE_TIMEOUT_SECS) {
+                self.info_message = None;
+            }
+        }
+
+        self.setup_watcher(ctx);
+        self.process_watcher_events();
+        self.process_async_results();
+        self.handle_input(ctx);
+
+        // Handle files dropped from external sources
+        ctx.input(|i| {
+            if !i.raw.dropped_files.is_empty() {
+                self.handle_dropped_files(&i.raw.dropped_files);
+            }
+        });
+
+        if self.mode == AppMode::Filtering {
+            let old_len = self.visible_entries.len();
+            self.apply_filter();
+            if self.visible_entries.len() != old_len {
+                self.last_selection_change = Instant::now();
+            }
+        }
+
+        let next_navigation = std::cell::RefCell::new(None);
+        let next_selection = std::cell::RefCell::new(None);
+        let pending_selection = std::cell::RefCell::new(None);
+        let context_action = std::cell::RefCell::new(None::<Box<dyn FnOnce(&mut Self)>>);
+
+        egui::TopBottomPanel::top("top_panel").show(ctx, |ui| {
+            ui.add_space(4.0);
+            ui.horizontal(|ui| {
+                // History Controls (fixed)
+                if ui.button("⬅").on_hover_text("Back (Alt+Left)").clicked() {
+                    self.navigate_back();
+                }
+                if ui
+                    .button("➡")
+                    .on_hover_text("Forward (Alt+Right)")
+                    .clicked()
+                {
+                    self.navigate_forward();
+                }
+                if ui.button("⬆").on_hover_text("Up (Backspace)").clicked() {
+                    self.navigate_up();
+                }
+                ui.add_space(10.0);
+
+                // Breadcrumbs (scrollable) - reserve space for right controls
+                let breadcrumb_width = ui.available_width() - 180.0;
+                egui::ScrollArea::horizontal()
+                    .id_salt("breadcrumbs")
+                    .max_width(breadcrumb_width)
+                    .show(ui, |ui| {
+                        ui.horizontal(|ui| {
+                            let components: Vec<_> = self.current_path.components().collect();
+                            let mut path_acc = PathBuf::new();
+                            for component in components {
+                                path_acc.push(component);
+                                let name = component.as_os_str().to_string_lossy();
+                                let label = if name.is_empty() { "/" } else { &name };
+                                if ui.button(label).clicked() {
+                                    *next_navigation.borrow_mut() = Some(path_acc.clone());
+                                }
+                                ui.label(">");
+                            }
+                        });
+                    });
+
+                // Right controls in remaining space
+                ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
+                    if ui.checkbox(&mut self.show_hidden, "Hidden (.)").changed() {
+                        self.request_refresh();
+                    }
+
+                    // Theme toggle
+                    let theme_icon = match self.theme {
+                        Theme::Light => "🌙",
+                        Theme::Dark => "☀",
+                    };
+                    if ui
+                        .button(theme_icon)
+                        .on_hover_text("Toggle theme")
+                        .clicked()
+                    {
+                        self.theme = match self.theme {
+                            Theme::Light => Theme::Dark,
+                            Theme::Dark => Theme::Light,
+                        };
+                    }
+
+                    if ui.button("?").clicked() {
+                        self.mode = AppMode::Help;
+                    }
+
+                    // Mode Indicator
+                    match &self.mode {
+                        AppMode::Normal => {
+                            ui.label("NORMAL");
+                        }
+                        AppMode::Visual => {
+                            ui.colored_label(egui::Color32::LIGHT_BLUE, "VISUAL");
+                        }
+                        AppMode::Filtering => {
+                            ui.colored_label(egui::Color32::YELLOW, "FILTER");
+                        }
+                        AppMode::Command => {
+                            ui.colored_label(egui::Color32::RED, "COMMAND");
+                        }
+                        AppMode::Help => {
+                            ui.colored_label(egui::Color32::GREEN, "HELP");
+                        }
+                        AppMode::Rename => {
+                            ui.colored_label(egui::Color32::ORANGE, "RENAME");
+                        }
+                        AppMode::DeleteConfirm => {
+                            ui.colored_label(egui::Color32::RED, "CONFIRM DELETE?");
+                        }
+                        AppMode::SearchInput => {
+                            ui.colored_label(egui::Color32::LIGHT_BLUE, "SEARCH");
+                        }
+                        AppMode::SearchResults { results, .. } => {
+                            ui.colored_label(
+                                egui::Color32::LIGHT_BLUE,
+                                format!("SEARCH ({} results)", results.len()),
+                            );
+                        }
+                    }
+                });
+            });
+            ui.add_space(4.0);
+        });
+
+        egui::TopBottomPanel::bottom("bottom_panel").show(ctx, |ui| {
+            ui.horizontal(|ui| {
+                // Item counts
+                ui.label(format!(
+                    "{}/{} items",
+                    self.visible_entries.len(),
+                    self.all_entries.len()
+                ));
+                
+                // Show current selected file info
+                if let Some(idx) = self.selected_index {
+                    if let Some(entry) = self.visible_entries.get(idx) {
+                        ui.separator();
+                        let type_str = if entry.is_dir { "dir" } else { "file" };
+                        ui.label(format!(
+                            "{}: {}",
+                            type_str,
+                            bytesize::ByteSize(entry.size)
+                        ));
+                    }
+                }
+                
+                // Show current path
+                ui.separator();
+                style::truncated_label(ui, format!("{}", self.current_path.display()));
+
+                if self.is_loading {
+                    ui.spinner();
+                }
+
+                if let Some((msg, _)) = &self.info_message {
+                    ui.colored_label(egui::Color32::GREEN, msg);
+                }
+                if let Some((err, _)) = &self.error_message {
+                    ui.colored_label(egui::Color32::RED, format!(" | {}", err));
+                }
+
+                if !self.multi_selection.is_empty() {
+                    ui.separator();
+                    // Calculate total size of selected files
+                    let total_size: u64 = self.all_entries.iter()
+                        .filter(|e| self.multi_selection.contains(&e.path))
+                        .map(|e| e.size)
+                        .sum();
+                    ui.colored_label(
+                        egui::Color32::LIGHT_BLUE,
+                        format!("{} selected ({})", 
+                            self.multi_selection.len(),
+                            bytesize::ByteSize(total_size)
+                        ),
+                    );
+                }
+            });
+        });
+
+        // Search Results View
+        if let AppMode::SearchResults {
+            ref query,
+            ref results,
+            selected_index,
+        } = self.mode
+        {
+            egui::CentralPanel::default().show(ctx, |ui| {
+                ui.add_space(4.0);
+                ui.horizontal(|ui| {
+                    ui.heading(format!("Search Results: \"{}\"", query));
+                    ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
+                        ui.label(format!("{} matches", results.len()));
+                    });
+                });
+                ui.separator();
+                ui.add_space(4.0);
+
+                ui.columns(2, |columns| {
+                    // Left column: Results list
+                    columns[0].vertical(|ui| {
+                        ui.heading("Matches");
+                        ui.separator();
+                        egui::ScrollArea::vertical()
+                            .id_salt("search_results_scroll")
+                            .auto_shrink([false, false])
+                            .max_height(ui.available_height())
+                            .show(ui, |ui| {
+                                ui.set_max_width(ui.available_width());
+                                use egui_extras::{Column, TableBuilder};
+                                let mut table = TableBuilder::new(ui)
+                                    .striped(true)
+                                    .resizable(false)
+                                    .cell_layout(egui::Layout::left_to_right(egui::Align::Center))
+                                    .column(Column::remainder().clip(true));
+
+                                if !results.is_empty() && selected_index < results.len() {
+                                    table = table
+                                        .scroll_to_row(selected_index, Some(egui::Align::Center));
+                                }
+
+                                table.body(|body| {
+                                    body.rows(40.0, results.len(), |mut row| {
+                                        let row_index = row.index();
+                                        let result = &results[row_index];
+                                        let is_selected = selected_index == row_index;
+
+                                        if is_selected {
+                                            row.set_selected(true);
+                                        }
+
+                                        row.col(|ui| {
+                                            ui.vertical(|ui| {
+                                                let file_label = format!(
+                                                    "{}:{}",
+                                                    result.file_name, result.line_number
+                                                );
+                                                let text = if is_selected {
+                                                    egui::RichText::new(&file_label).color(
+                                                        egui::Color32::from_rgb(100, 200, 255),
+                                                    )
+                                                } else {
+                                                    egui::RichText::new(&file_label)
+                                                };
+                                                ui.label(text);
+
+                                                // Show line content preview (truncated)
+                                                let preview = if result.line_content.len() > 60 {
+                                                    format!("{}...", &result.line_content[..60])
+                                                } else {
+                                                    result.line_content.clone()
+                                                };
+                                                ui.label(
+                                                    egui::RichText::new(preview)
+                                                        .size(10.0)
+                                                        .color(egui::Color32::GRAY),
+                                                );
+                                            });
+                                        });
+                                    });
+                                });
+                            });
+                    });
+
+                    // Right column: Preview
+                    columns[1].vertical(|ui| {
+                        ui.heading("Preview");
+                        ui.separator();
+
+                        if let Some(result) = results.get(selected_index) {
+                            ui.label(egui::RichText::new(&result.file_name).strong());
+                            ui.separator();
+
+                            // Show context around the match
+                            egui::ScrollArea::vertical()
+                                .id_salt("search_preview_scroll")
+                                .auto_shrink([false, false])
+                                .max_height(ui.available_height())
+                                .show(ui, |ui| {
+                                    ui.set_max_width(ui.available_width());
+                                    ui.horizontal(|ui| {
+                                        ui.label(format!("Line {}:", result.line_number));
+                                        ui.label(egui::RichText::new(&result.line_content).code());
+                                    });
+
+                                    ui.add_space(10.0);
+                                    ui.label("Full file path:");
+                                    ui.label(
+                                        egui::RichText::new(result.file_path.display().to_string())
+                                            .code(),
+                                    );
+
+                                    ui.add_space(10.0);
+                                    ui.horizontal(|ui| {
+                                        ui.label("Press");
+                                        ui.label(egui::RichText::new("Enter").strong());
+                                        ui.label("to open file,");
+                                        ui.label(egui::RichText::new("n/N").strong());
+                                        ui.label("for next/previous,");
+                                        ui.label(egui::RichText::new("Esc").strong());
+                                        ui.label("to return");
+                                    });
+                                });
+                        }
+                    });
+                });
+            });
+        } else {
+            // Normal file browser view
+            // Visual feedback for drag and drop
+            let is_being_dragged_over = ctx.input(|i| !i.raw.hovered_files.is_empty());
+
+            egui::CentralPanel::default().show(ctx, |ui| {
+                // Show drop zone overlay when files are being dragged over
+                if is_being_dragged_over {
+                    let painter = ui.painter();
+                    let rect = ui.available_rect_before_wrap();
+                    painter.rect_stroke(
+                        rect,
+                        5.0,
+                        egui::Stroke::new(2.0, egui::Color32::from_rgb(100, 200, 255)),
+                        egui::epaint::StrokeKind::Outside,
+                    );
+                    ui.label(
+                        egui::RichText::new("📁 Drop files here to copy them to this directory")
+                            .size(16.0)
+                            .color(egui::Color32::from_rgb(100, 200, 255)),
+                    );
+                }
+                // Help Modal
+                if self.mode == AppMode::Help {
+                    egui::Window::new("Help")
+                        .collapsible(false)
+                        .resizable(false)
+                        .anchor(egui::Align2::CENTER_CENTER, [0.0, 0.0])
+                        .default_width(style::modal_width(ctx))
+                        .show(ctx, |ui| {
+                            ui.set_max_height(style::modal_max_height(ctx));
+                            egui::ScrollArea::vertical().show(ui, |ui| {
+                                ui.heading("Key Bindings");
+                                ui.separator();
+                                egui::Grid::new("help_grid").striped(true).show(ui, |ui| {
+                                    ui.label("j / Down");
+                                    ui.label("Next Item");
+                                    ui.end_row();
+                                    ui.label("k / Up");
+                                    ui.label("Previous Item");
+                                    ui.end_row();
+                                    ui.label("h / Left Arrow / Backspace");
+                                    ui.label("Go to Parent");
+                                    ui.end_row();
+                                    ui.label("l / Right Arrow");
+                                    ui.label("Enter Directory");
+                                    ui.end_row();
+                                    ui.label("Enter");
+                                    ui.label("Open File / Enter Dir");
+                                    ui.end_row();
+                                    ui.label("gg / G");
+                                    ui.label("Top / Bottom");
+                                    ui.end_row();
+                                    ui.label("Alt + Arrows");
+                                    ui.label("History");
+                                    ui.end_row();
+                                    ui.label(".");
+                                    ui.label("Toggle Hidden");
+                                    ui.end_row();
+                                    ui.label("/");
+                                    ui.label("Filter Mode");
+                                    ui.end_row();
+                                    ui.label("S (Shift+s)");
+                                    ui.label("Content Search");
+                                    ui.end_row();
+                                    ui.label(":");
+                                    ui.label("Command Mode");
+                                    ui.end_row();
+                                    ui.label("v");
+                                    ui.label("Visual Select Mode");
+                                    ui.end_row();
+                                    ui.label("y / x / p");
+                                    ui.label("Copy / Cut / Paste");
+                                    ui.end_row();
+                                    ui.label("d / r");
+                                    ui.label("Delete / Rename");
+                                    ui.end_row();
+                                    ui.label("?");
+                                    ui.label("Toggle Help");
+                                    ui.end_row();
+                                    ui.label("Shift+V");
+                                    ui.label("Visual Mode (Select All)");
+                                    ui.end_row();
+                                    ui.label("Ctrl+A");
+                                    ui.label("Select All Items");
+                                    ui.end_row();
+                                    ui.label("Space");
+                                    ui.label("Toggle Selection");
+                                    ui.end_row();
+                                });
+                                ui.add_space(10.0);
+                                if ui.button("Close").clicked() {
+                                    self.mode = AppMode::Normal;
+                                }
+                            });
+                        });
+                }
+
+                // Search Input Modal
+                if self.mode == AppMode::SearchInput {
+                    egui::Window::new("Content Search")
+                        .collapsible(false)
+                        .resizable(false)
+                        .anchor(egui::Align2::CENTER_CENTER, [0.0, 0.0])
+                        .default_width(style::modal_width(ctx))
+                        .show(ctx, |ui| {
+                            ui.set_max_height(style::modal_max_height(ctx));
+                            egui::ScrollArea::vertical().show(ui, |ui| {
+                                ui.label("Search for content in files:");
+                                ui.add_space(5.0);
+
+                                let response = ui.text_edit_singleline(&mut self.search_query);
+                                if self.focus_input {
+                                    response.request_focus();
+                                    self.focus_input = false;
+                                }
+
+                                ui.add_space(10.0);
+                                ui.label("Options:");
+                                ui.checkbox(
+                                    &mut self.search_options.case_sensitive,
+                                    "Case sensitive",
+                                );
+                                ui.checkbox(&mut self.search_options.use_regex, "Use regex");
+                                ui.checkbox(
+                                    &mut self.search_options.search_hidden,
+                                    "Search hidden files",
+                                );
+                                ui.checkbox(&mut self.search_options.search_pdfs, "Search PDFs");
+                                ui.checkbox(
+                                    &mut self.search_options.search_archives,
+                                    "Search archives",
+                                );
+
+                                ui.add_space(10.0);
+                                ui.horizontal(|ui| {
+                                    if ui.button("Search").clicked()
+                                        && !self.search_query.is_empty()
+                                    {
+                                        self.search_in_progress = true;
+                                        self.search_file_count = 0;
+                                        let _ = self.command_tx.send(IoCommand::SearchContent {
+                                            query: self.search_query.clone(),
+                                            root_path: self.current_path.clone(),
+                                            options: self.search_options.clone(),
+                                        });
+                                        self.mode = AppMode::Normal;
+                                    }
+                                    if ui.button("Cancel").clicked() {
+                                        self.mode = AppMode::Normal;
+                                    }
+                                });
+
+                                if self.search_in_progress {
+                                    ui.add_space(10.0);
+                                    ui.horizontal(|ui| {
+                                        ui.spinner();
+                                        ui.label(format!(
+                                            "Searching... ({} files)",
+                                            self.search_file_count
+                                        ));
+                                    });
+                                }
+                            });
+                        });
+                }
+
+                // Command/Filter/Rename Input Modal
+                if matches!(
+                    self.mode,
+                    AppMode::Command | AppMode::Filtering | AppMode::Rename
+                ) {
+                    egui::Area::new("input_popup".into())
+                        .anchor(egui::Align2::CENTER_TOP, [0.0, 50.0])
+                        .order(egui::Order::Foreground)
+                        .show(ctx, |ui| {
+                            egui::Frame::popup(ui.style()).show(ui, |ui| {
+                                ui.set_min_width(400.0);
+                                let prefix = match self.mode {
+                                    AppMode::Rename => "Rename:",
+                                    AppMode::Filtering => "/",
+                                    _ => ":",
+                                };
+                                ui.horizontal(|ui| {
+                                    ui.label(prefix);
+                                    let response =
+                                        ui.text_edit_singleline(&mut self.command_buffer);
+                                    if self.focus_input {
+                                        response.request_focus();
+                                        self.focus_input = false;
+                                    }
+                                });
+                            });
+                        });
+                }
+
+                // Strip-based layout with three panes and dividers
+                use egui_extras::{Size, StripBuilder};
+                StripBuilder::new(ui)
+                    .size(Size::exact(self.panel_widths[0]).at_least(style::PARENT_MIN))
+                    .size(Size::exact(style::DIVIDER_WIDTH))
+                    .size(Size::remainder())
+                    .size(Size::exact(style::DIVIDER_WIDTH))
+                    .size(Size::exact(self.panel_widths[1]).at_least(style::PREVIEW_MIN))
+                    .horizontal(|mut strip| {
+                        strip.cell(|ui| self.render_parent_pane(ui, &next_navigation));
+                        strip.cell(|ui| self.render_divider(ui, 0));
+                        strip.cell(|ui| {
+                            self.render_current_pane(
+                                ui,
+                                &next_navigation,
+                                &next_selection,
+                                &context_action,
+                                ctx,
+                            )
+                        });
+                        strip.cell(|ui| self.render_divider(ui, 1));
+                        strip.cell(|ui| {
+                            ui.add_space(4.0);
+                            ui.vertical_centered(|ui| {
+                                ui.heading("Preview");
+                            });
+                            ui.separator();
+                            self.render_preview(ui, &next_navigation, &pending_selection);
+                        });
+                    });
+            });
+        } // End of else block for normal file browser view
+
+        if let Some(idx) = next_selection.into_inner() {
+            self.selected_index = Some(idx);
+        }
+        if let Some(pending) = pending_selection.into_inner() {
+            self.pending_selection_path = Some(pending);
+        }
+        if let Some(path) = next_navigation.into_inner() {
+            self.navigate_to(path);
+        }
+        if let Some(action) = context_action.into_inner() {
+            action(self);
+        }
+    }
+}
+
