+++ src/input.rs
@@ -1,0 +1,3 @@
+// Input handling methods for Heike
+// Keyboard and mouse input processing
+// Will be extracted from app.rs in future refactoring
