+++ src/io/mod.rs
@@ -1,0 +1,6 @@
+pub mod directory;
+pub mod search;
+pub mod worker;
+
+pub use directory::{fuzzy_match, is_likely_binary, read_directory};
+pub use worker::{spawn_worker, IoCommand, IoResult};
