+++ src/state/clipboard.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Copy, PartialEq)]
+pub enum ClipboardOp {
+    Copy,
+    Cut,
+}
