+++ src/state/mod.rs
@@ -1,0 +1,7 @@
+pub mod clipboard;
+pub mod mode;
+pub mod search;
+
+pub use clipboard::ClipboardOp;
+pub use mode::AppMode;
+pub use search::{SearchOptions, SearchResult};
