+++ src/view/mod.rs
@@ -1,0 +1,3 @@
+pub mod modals;
+pub mod panels;
+pub mod preview;
