+++ src/view/modals.rs
@@ -1,0 +1,3 @@
+// Modal rendering functions (Help, Rename, Delete, Search)
+// Will be kept in app.rs update() method for now due to tight coupling
+// This module can be used for future extraction
