+++ src/view/panels.rs
@@ -1,0 +1,3 @@
+// Panel rendering functions will be kept in app.rs for now
+// due to tight coupling with Heike struct and update loop
+// This module can be used for future extraction of panel-specific helpers
